Move ButtonActionModel to separate file
