+++ CTIntegrationTests/Assets/Scripts/Utilities/ButtonActionModel.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d8469220b121249cebeb5b829cc35e4e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
